@@ -22,27 +22,19 @@
   "author": "<bluebill1049@hotmail.com>",
   "license": "MIT",
   "devDependencies": {
-    "@types/enzyme": "^3.9.0",
-    "@types/enzyme-adapter-react-16": "^1.0.5",
-    "@types/jest": "^24.0.17",
-    "@types/react": "^16.8.8",
-    "@types/react-test-renderer": "^16.8.1",
-    "coveralls": "^3.0.3",
-    "enzyme": "^3.9.0",
-    "enzyme-adapter-react-16": "^1.9.1",
-    "jest": "24.7.1",
+    "@types/lodash": "^4.17.14",
+    "@types/react": "^19.0.4",
+    "coveralls": "^3.1.1",
     "npm-run-all": "^4.1.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-test-renderer": "^16.8.3",
-    "rimraf": "^2.6.3",
-    "rollup": "^1.6.0",
-    "rollup-plugin-peer-deps-external": "^2.2.2",
-    "rollup-plugin-typescript2": "^0.19.2",
-    "ts-jest": "^24.0.0",
-    "typescript": "^3.3.3333",
-    "uglify-es": "^3.3.9",
-    "yarn": "^1.22.10"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "rimraf": "^6.0.1",
+    "rollup": "^4.30.1",
+    "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-typescript2": "^0.36.0",
+    "ts-jest": "^29.2.5",
+    "typescript": "^5.7.3",
+    "uglify-es": "^3.3.9"
   },
   "peerDependencies": {
     "little-state-machine": "^5.0.0",
@@ -50,14 +42,10 @@
     "react-dom": "^16.8.0 || ^17 || ^18"
   },
   "dependencies": {
-    "little-state-machine": "^5.0.0",
-    "lodash.clonedeep": "^4.5.0",
-    "lodash.isempty": "^4.4.0",
-    "lodash.isequal": "^4.5.0",
-    "lodash.isobject": "^3.0.2",
-    "lodash.transform": "^4.6.0",
-    "react-json-view": "^1.19.1",
-    "react-simple-animate": "^3.3.0"
+    "little-state-machine": "^5.0.1",
+    "lodash": "^4.17.21",
+    "react-json-view": "^1.21.3",
+    "react-simple-animate": "^3.5.3"
   },
   "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
 }
