@@ -1,8 +1,8 @@
-import React from 'react';
-import './App.css';
+import './App.css'
+
 import Form from './Form';
 import yourDetails from './states/yourDetails';
-import { StateMachineProvider, createStore } from 'little-state-machine';
+import { createStore } from 'little-state-machine';
 import { DevTool } from './src';
 
 createStore(
@@ -11,16 +11,16 @@ createStore(
   },
 );
 
-const App: React.FC = () => {
+function App() {
   return (
-    <StateMachineProvider>
+    <>
       <DevTool />
       <div className="App">
         <h1>Little State Machine</h1>
         <Form />
       </div>
-    </StateMachineProvider>
+    </>
   );
-};
+}
 
-export default App;
+export default App
